+++ js/data.js
@@ -1,0 +1,7 @@
+import {fotosFunctions} from'./util.js';
+const functions = fotosFunctions();
+// eslint-disable-next-line
+console.log(functions);
+
+
+
